+++ js/main.js
@@ -1,5 +1,5 @@
 import { getData } from './api.js';
-import { drawPhotos } from './draw.js';
+import { drawPhotos } from './drawPictures.js';
 import { activateFilters } from './fileFilter.js';
 import './form.js';
 import { onFail } from './util.js';
